Feat: Improve folder structure for scalability and maintainability
- Now crud.py and schemas.py will be allocated in separated folders with its model's name
- Password and jwt specific functions are now in separated files in security folder
