+++ app/core/security/handle_jwt.py
@@ -1,10 +1,10 @@
 import jwt
 from fastapi.security import OAuth2PasswordBearer
-from passlib.context import CryptContext
-from datetime import datetime, timedelta
-
 from fastapi import HTTPException, Request
 
+from datetime import datetime, timedelta
+
+from .passwords import ALGORITHM, SECRET_KEY
 
 class OAuth2PasswordCookie(OAuth2PasswordBearer):
     """OAuth2 password flow with token in a httpOnly cookie.
@@ -44,17 +44,3 @@ def create_access_token(*, data: dict, expires_delta: timedelta = None):
     to_encode.update({"exp": expire})
     encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
     return encoded_jwt
-
-# PASSWORD RELATED
-pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
-
-SECRET_KEY = "super_secret"
-ALGORITHM = "HS256"
-ACCESS_TOKEN_EXPIRE_MINUTES = 30
-
-def get_password_hash(password: str) -> str:
-    return pwd_context.hash(password)
-
-
-def verify_password(plain_password: str, hashed_password: str) -> bool:
-    return pwd_context.verify(plain_password, hashed_password)
